+++ docs-next/src/components/Layout.tsx
@@ -1,11 +1,11 @@
-import React, { ReactElement } from 'react';
+import React, { ReactNode } from 'react';
 import { Box } from '@marigold/components';
 
 import { NavigationMenu } from '~/navigation.utils';
 import { NavigationLayout, Version } from '~/components';
 
 export interface LayoutProps {
-  children: ReactElement;
+  children: ReactNode;
   navigation: NavigationMenu;
 }
 
@@ -22,7 +22,7 @@ export const Layout = ({ navigation, children }: LayoutProps) => {
       >
         <NavigationLayout navigation={navigation}></NavigationLayout>
 
-        <Box>{children}</Box>
+        <div>{children}</div>
       </Box>
       <Version />
     </>
