+++ docs-next/src/theme/components/Headline.style.ts
@@ -34,6 +34,7 @@ export const Headline: Theme['components']['Headline'] = {
     },
     'level-3': {
       my: 'none',
+      pb: 'medium-1',
       fontSize: 'small-1',
       fontWeight: 'medium',
       lineHeight: 'small-1',
